Clean up structure of io/console and avoid stty on Windows.
Fixes #3989.

* Restructure the different impls of console into their own files.
* Always use stubbed version on Windows.
* Cascade from native to stty to stubbed on other platforms.
