+++ lib/ruby/stdlib/io/console/bsd_console.rb
